+++ src/Plutarch/Extra/Const.hs
@@ -15,8 +15,6 @@ module Plutarch.Extra.Const (
 
 import Data.Kind (Type)
 import GHC.Generics (Generic)
-import Generics.SOP (Top)
-import qualified Generics.SOP as SOP
 import Plutarch (
     DerivePlutusType (..),
     PlutusType,
@@ -36,6 +34,7 @@ import Plutarch.Extra.Boring (PBoring (pboring))
 import Plutarch.Extra.Functor (
     PBifunctor (PSubcategoryLeft, PSubcategoryRight, pbimap, psecond),
     PFunctor (PSubcategory, pfmap),
+    Plut,
  )
 import Plutarch.Extra.TermCont (pmatchC)
 import Plutarch.Integer (PIntegral)
@@ -55,8 +54,6 @@ newtype PConst (a :: S -> Type) (b :: S -> Type) (s :: S)
         )
     deriving anyclass
         ( -- | @since 1.0.0
-          SOP.Generic
-        , -- | @since 1.0.0
           PlutusType
         )
 
@@ -85,15 +82,15 @@ deriving anyclass instance (PNum a) => PNum (PConst a b)
 -- | @since 1.0.0
 deriving anyclass instance (PShow a) => PShow (PConst a b)
 
--- | @since 1.0.0
+-- | @since 3.1.0
 instance PFunctor (PConst a) where
-    type PSubcategory (PConst a) = Top
+    type PSubcategory (PConst a) = Plut
     pfmap = psecond
 
--- | @since 1.0.0
+-- | @since 3.1.0
 instance PBifunctor PConst where
-    type PSubcategoryLeft PConst = Top
-    type PSubcategoryRight PConst = Top
+    type PSubcategoryLeft PConst = Plut
+    type PSubcategoryRight PConst = Plut
     pbimap = phoistAcyclic $
         plam $ \f _ t -> unTermCont $ do
             PConst tx <- pmatchC t