+++ src/Plutarch/Extra/Monoid.hs
@@ -1,5 +1,3 @@
-{-# LANGUAGE DeriveAnyClass #-}
-{-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-}
 
@@ -13,7 +11,6 @@ module Plutarch.Extra.Monoid (
 import Control.Composition ((.*))
 import Data.Function (on)
 import GHC.Generics (Generic)
-import qualified Generics.SOP as SOP
 import Plutarch (
     DerivePlutusType (DPTStrat),
     PlutusType,
@@ -38,9 +35,7 @@ newtype PAll (s :: S) = PAll (Term s PBool)
           Generic
         )
     deriving anyclass
-        ( -- | @since 1.4.0
-          SOP.Generic
-        , -- | @since 1.3.0
+        ( -- | @since 1.3.0
           PlutusType
         , -- | @since 1.3.0
           PIsData
@@ -70,9 +65,7 @@ newtype PAny (s :: S) = PAny (Term s PBool)
           Generic
         )
     deriving anyclass
-        ( -- | @since 1.4.0
-          SOP.Generic
-        , -- | @since 1.3.0
+        ( -- | @since 1.3.0
           PlutusType
         , -- | @since 1.3.0
           PIsData