+++ src/Plutarch/Extra/Profunctor.hs
@@ -5,9 +5,9 @@ module Plutarch.Extra.Profunctor (
 ) where
 
 import Data.Kind (Constraint, Type)
-import Generics.SOP (Top)
 import Plutarch (S, Term, phoistAcyclic, plam, (#), type (:-->))
 import Plutarch.Extra.Function (pidentity)
+import Plutarch.Extra.Functor (Plut)
 
 -- | @since 1.0.0
 class PProfunctor (p :: (S -> Type) -> (S -> Type) -> S -> Type) where
@@ -40,8 +40,8 @@ class PProfunctor (p :: (S -> Type) -> (S -> Type) -> S -> Type) where
         Term s ((c :--> d) :--> p a c :--> p a d)
     prmap = phoistAcyclic $ plam $ \g p -> pdimap # pidentity # g # p
 
--- | @since 1.0.0
+-- | @since 3.1.0
 instance PProfunctor (:-->) where
-    type PContraSubcategory (:-->) = Top
-    type PCoSubcategory (:-->) = Top
+    type PContraSubcategory (:-->) = Plut
+    type PCoSubcategory (:-->) = Plut
     pdimap = phoistAcyclic $ plam $ \ab cd bc -> plam $ \x -> cd # (bc # (ab # x))