+++ src/Plutarch/Extra/Star.hs
@@ -13,7 +13,6 @@ module Plutarch.Extra.Star (
 
 import Data.Kind (Type)
 import GHC.Generics (Generic)
-import Generics.SOP (Top)
 import Plutarch (
     DerivePlutusType (DPTStrat),
     PlutusType,
@@ -31,7 +30,7 @@ import Plutarch (
 import Plutarch.Extra.Applicative (PApplicative (ppure), PApply (pliftA2))
 import Plutarch.Extra.Bind (PBind ((#>>=)))
 import Plutarch.Extra.Category (PCategory (pidentity), PSemigroupoid ((#>>>)))
-import Plutarch.Extra.Functor (PFunctor (PSubcategory, pfmap))
+import Plutarch.Extra.Functor (PFunctor (PSubcategory, pfmap), Plut)
 import Plutarch.Extra.Profunctor (PProfunctor (PCoSubcategory, PContraSubcategory, pdimap))
 import Plutarch.Extra.TermCont (pmatchC)
 
@@ -66,10 +65,10 @@ instance DerivePlutusType (PStar f a b) where
 {- | If @f@ is at least a 'PFunctor', we can pre-process and post-process work
  done in @PStar f@ using pure functions.
 
- @since 3.0.1
+ @since 3.1.0
 -}
 instance (PFunctor f) => PProfunctor (PStar f) where
-    type PContraSubcategory (PStar f) = Top
+    type PContraSubcategory (PStar f) = Plut
     type PCoSubcategory (PStar f) = PSubcategory f
     pdimap = phoistAcyclic $
         plam $ \into outOf xs -> unTermCont $ do