+++ src/Plutarch/Extra/Sum.hs
@@ -10,8 +10,6 @@ module Plutarch.Extra.Sum (
 
 import Data.Kind (Type)
 import GHC.Generics (Generic)
-import Generics.SOP (Top)
-import qualified Generics.SOP as SOP
 import Plutarch (
     DerivePlutusType (..),
     PlutusType,
@@ -29,7 +27,7 @@ import Plutarch.Builtin (PIsData)
 import Plutarch.Extra.Applicative (PApplicative (ppure), PApply (pliftA2))
 import Plutarch.Extra.Boring (PBoring (pboring))
 import Plutarch.Extra.Comonad (PComonad (pextract), PExtend (pextend))
-import Plutarch.Extra.Functor (PFunctor (PSubcategory, pfmap))
+import Plutarch.Extra.Functor (PFunctor (PSubcategory, pfmap), Plut)
 import Plutarch.Extra.TermCont (pmatchC)
 import Plutarch.Integer (PIntegral)
 import Plutarch.Num (PNum)
@@ -47,8 +45,6 @@ newtype PSum (a :: S -> Type) (s :: S)
         )
     deriving anyclass
         ( -- | @since 1.0.0
-          SOP.Generic
-        , -- | @since 1.0.0
           PlutusType
         )
 
@@ -94,9 +90,9 @@ instance
 -- | @since 1.0.0
 deriving anyclass instance (PShow a) => PShow (PSum a)
 
--- | @since 1.0.0
+-- | @since 3.1.0
 instance PFunctor PSum where
-    type PSubcategory PSum = Top
+    type PSubcategory PSum = Plut
     pfmap = phoistAcyclic $
         plam $ \f t -> unTermCont $ do
             PSum t' <- pmatchC t