+++ src/Plutarch/Extra/These.hs
@@ -19,7 +19,6 @@ module Plutarch.Extra.These (
 
 import Data.Kind (Type)
 import GHC.Generics (Generic)
-import Generics.SOP (Top)
 import Plutarch.Bool (PBool (PFalse, PTrue), PEq, POrd, PPartialOrd (..))
 import Plutarch.Builtin (PAsData, PIsData, pdata, pfromData)
 import Plutarch.DataRepr (
@@ -35,6 +34,7 @@ import Plutarch.Extra.Boring (pboring)
 import Plutarch.Extra.Functor (
     PBifunctor (PSubcategoryLeft, PSubcategoryRight, pbimap, psecond),
     PFunctor (PSubcategory, pfmap),
+    Plut,
  )
 import Plutarch.Extra.TermCont (pletC, pmatchC)
 import Plutarch.Extra.Traversable (PTraversable (ptraverse, ptraverse_))
@@ -101,15 +101,15 @@ instance (POrd a, POrd b) => PPartialOrd (PThese a b) where
 -- | @since 1.0.0
 deriving anyclass instance (PShow a, PShow b) => PShow (PThese a b)
 
--- | @since 1.0.0
+-- | @since 3.1.0
 instance PFunctor (PThese a) where
-    type PSubcategory (PThese a) = Top
+    type PSubcategory (PThese a) = Plut
     pfmap = psecond
 
--- | @since 1.0.0
+-- | @since 3.1.0
 instance PBifunctor PThese where
-    type PSubcategoryLeft PThese = Top
-    type PSubcategoryRight PThese = Top
+    type PSubcategoryLeft PThese = Plut
+    type PSubcategoryRight PThese = Plut
     pbimap = phoistAcyclic $
         plam $ \f g t -> unTermCont $ do
             t' <- pmatchC t